Ekona: Add Vector3 class to mathematics namespace with refactoring NitroTextureCompression
