+++ Ekona/Ekona.csproj
@@ -109,6 +109,8 @@
       <DependentUpon>PaletteControl.cs</DependentUpon>
     </Compile>
     <Compile Include="Images\RawData.cs" />
+    <Compile Include="Mathematics\NvMath.cs" />
+    <Compile Include="Mathematics\Vector3.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -138,6 +140,7 @@
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
